pyproject.toml
[tool.poetry]
name = "disk-destiny"
version = "0.1.0"
description = ""
authors = ["DIMFLIX-OFFICIAL <dimflix.official@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
inquirer = "^3.4.0"
numpy = "^2.1.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
pypickle = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
